Refactor: Update login-signup modal and main css
