+++ src/main/resources/example.properties
@@ -21,6 +21,10 @@ spring.security.oauth2.client.registration.google.client-id=1232313223-3213.apps
 spring.security.oauth2.client.registration.google.client-secret=yoursecretkey
 spring.security.oauth2.client.registration.google.scope=profile,email
 
+#multipart files
+spring.servlet.multipart.max-file-size=10MB
+spring.servlet.multipart.max-request-size=10MB
+
 #aws
 aws.s3.endpoint.url=https://s3.us-east-1.amazonaws.com
 aws.s3.access.key=ACCESSKEY
