+++ src/main/resources/templates/error/403.html
@@ -12,7 +12,7 @@
             padding: 0;
         }
         body{
-            /*background: #233142;*/
+            background: #03b090;
 
         }
         .whistle{
@@ -54,9 +54,13 @@
             text-transform: uppercase;
         }
     </style>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+
 </head>
 
 <body>
+<div style="background-color: #03b090" class="alert alert-green text-center text-white" role="alert" id="timer"></div>
+
 <use>
     <svg version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" x="0px" y="0px" viewBox="0 0 1000 1000" enable-background="new 0 0 1000 1000" xml:space="preserve" class="whistle">
 <metadata> Svg Vector Icons : http://www.onlinewebfonts.com/icon </metadata>
@@ -66,5 +70,22 @@
 </use>
 <h1>403</h1>
 <h2>Access forbidden!</h2>
+<!-- Redirection Counter -->
+<script type="text/javascript">
+    let count = 10; // Timer
+    const redirect = "/"; // Target URL
+
+    const countDown = ()=> {
+        let timer = document.getElementById("timer"); // Timer ID
+        if (count > 0) {
+            count--;
+            timer.innerHTML = "This page will redirect in " + count + " seconds."; // Timer Message
+            setTimeout("countDown()", 1000);
+        } else {
+            window.location.href = redirect;
+        }
+    }
+    countDown()
+</script>
 </body>
 </html>
