--- src/main/resources/templates/login.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
-<head th:replace="partials/head :: head(title)"></head>
-<body>
-
-<header th:include="partials/navbar.html :: nav"></header>
-
-<!-- Content -->
-<div class="container">
-  <!--Grid row-->
-  <div class="row mt-5">
-    <!--Grid column-->
-    <div class="col-md-6 mb-5 mt-md-0 mt-5 white-text text-center text-md-left">
-      <h1 class="h1-responsive font-weight-bold wow fadeInLeft" data-wow-delay="0.3s">Sign up right now! </h1>
-      <hr class="hr-light wow fadeInLeft" data-wow-delay="0.3s">
-      <h6 class="mb-3 wow fadeInLeft" data-wow-delay="0.3s">Lorem ipsum dolor sit amet, consectetur adipisicing elit. Rem repellendus quasi fuga
-        nesciunt dolorum nulla magnam veniam sapiente, fugiat! Commodi sequi non animi ea
-        dolor molestiae, quisquam iste, maiores. Nulla.</h6>
-      <a class="btn btn-outline-white wow fadeInLeft" data-wow-delay="0.3s">Learn more</a>
-    </div>
-
-    <div class="col-md-6 col-xl-5 mb-4">
-      <!-- Login Form -->
-      <form th:action="@{/login}" class="text-center border border-light p-5" th:object="${user}">
-
-        <p class="h4 mb-4">Log in</p>
-
-        <!-- E-mail -->
-        <input type="email" id="loginEmail" class="form-control mb-4" placeholder="E-mail">
-
-        <!-- Password -->
-        <input type="password" id="loginPassword" class="form-control" placeholder="Password" aria-describedby="defaultRegisterFormPasswordHelpBlock">
-
-        <!-- Remember Me -->
-        <div class="custom-control custom-checkbox">
-          <input type="checkbox" class="custom-control-input" id="rememberMe">
-          <label class="custom-control-label" for="rememberMe">Remember me</label>
-        </div>
-
-        <!-- Sign up button -->
-        <button class="btn btn-info my-4 btn-block" type="submit">Sign in</button>
-
-        <!-- Social register -->
-        <p>or login in with:</p>
-        <a href="#" class="mx-2" role="button"><i class="fab fa-google"></i></a>
-        <a href="#" class="mx-2" role="button"><i class="fab fa-facebook-f light-blue-text"></i></a>
-        <a href="#" class="mx-2" role="button"><i class="fab fa-twitter light-blue-text"></i></a>
-
-        <!-- Terms of service -->
-      </form>
-      <!-- Default form register -->
-    </div>
-    <!--Grid column-->
-  </div>
-  <!--Grid row-->
-</div>
-<!-- Content -->
-
-<!--/. Sidebar navigation -->
-<!-- /Start your project here-->
-<!-- JQuery -->
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.4.1/jquery.min.js"></script>
-<!-- Bootstrap tooltips -->
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.4/umd/popper.min.js"></script>
-<!-- Bootstrap core JavaScript -->
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.4.1/js/bootstrap.min.js"></script>
-<!-- MDB core JavaScript -->
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/mdbootstrap/4.13.0/js/mdb.min.js"></script>
-</body>
-</html>
-
-
-
-<!--Todo: Form Doesn't work-->
