+++ src/main/resources/templates/partials/modals.html
@@ -6,6 +6,9 @@
                     <span aria-hidden="true">&times;</span>
                 </button>
                 <div id="regForm"></div>
+                <div class="text-center">
+                    <a class="login-signup-toggle text-decoration-none">Already a member? Log in</a>
+                </div>
 <!--                <form th:replace="users/forms :: register"></form>-->
             </div>
         </div>
@@ -20,6 +23,9 @@
                     <span aria-hidden="true">&times;</span>
                 </button>
                 <form th:replace="users/forms :: login"></form>
+                <div class="text-center">
+                    <a class="login-signup-toggle text-decoration-none">Not on ChowNow? Sign up</a>
+                </div>
             </div>
         </div>
     </div>
