+++ src/main/resources/templates/sessions/new.html
@@ -9,7 +9,7 @@
         <div class="col-md-3 mb-5 mt-md-0 mt-5 white-text text-center text-md-left">
         </div>
         <div class="col-md-6 mb-4">
-            <div class="card rounded-card">
+            <div class="card rounded-card bg-light">
                 <form th:replace="users/forms :: login"></form>
             </div>
         </div>
